+++ compiler-cli/src/new.rs
@@ -11,11 +11,11 @@ use strum::{Display, EnumString, EnumVariantNames};
 
 use crate::NewOptions;
 
-const GLEAM_STDLIB_VERSION: &str = "0.25";
-const GLEEUNIT_VERSION: &str = "0.7";
-const ERLANG_OTP_VERSION: &str = "25.1";
-const REBAR3_VERSION: &str = "25.1";
-const ELIXIR_VERSION: &str = "1.14.1";
+const GLEAM_STDLIB_VERSION: &str = "0.25.0";
+const GLEEUNIT_VERSION: &str = "0.7.2";
+const ERLANG_OTP_VERSION: &str = "25.2";
+const REBAR3_VERSION: &str = "3.20.0";
+const ELIXIR_VERSION: &str = "1.14.2";
 
 #[derive(Debug, Serialize, Deserialize, Display, EnumString, EnumVariantNames, Clone, Copy)]
 #[strum(serialize_all = "kebab_case")]
@@ -154,7 +154,7 @@ jobs:
   test:
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v3.0.0
+      - uses: actions/checkout@v3.2.0
       - uses: erlef/setup-beam@v1.15.0
         with:
           otp-version: "{}"